# perf(123pan): optimize rate limiting