perf(123pan): optimize rate limiting

- eliminating fixed 200 ms delay in getFiles to prevent thread starvation
Under a concurrency of 5 requests, performance improvement is as follows:

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `before` | 205.099 ± 14.682 | 188.500 | 227.832 | 1.13 ± 0.08 |
| `after` | 182.253 ± 0.188 | 182.112 | 182.488 | 1.00 |

![hyperfine benchmark result](https://github.com/user-attachments/assets/071170d7-c34e-488e-877b-95e01434ea95)

<details><summary>Benchmark Commands</summary>

```
31812 (seiuneko@R9000P) /w
❯❯❯ cat benchmark
#!/usr/bin/zsh

export user=admin pass='D8@Sj&VDXKCM%47s'
# seq 0 255 | xargs printf "%02x\n" > dir_list
xargs -a dir_list -I _ -P 5 curl -u $user:$pass -X PROPFIND -H 'Depth: 1' -o /dev/null -s http://127.0.233.1:5244/dav/123pan/restic-repo/data/_

31804 (seiuneko@R9000P) /w
❯❯❯ hyperfine --shell zsh --runs 5 --prepare './alist-before --data /w/alist-run server 2>/dev/null &; sleep 1' ./benchmark --conclude 'pkill alist; sleep 60' --prepare './alist-after --data /w/alist-run server 2>/dev/null &; sleep 1' ./benchmark --conclude 'pkill alist; sleep 60' --export-markdown result.md --command-name before --command-name after
Benchmark 1: before
  Time (mean ± σ):     205.099 s ± 14.682 s    [User: 0.574 s, System: 0.650 s]
  Range (min … max):   188.500 s … 227.832 s    5 runs

Benchmark 2: after
  Time (mean ± σ):     182.253 s ±  0.188 s    [User: 0.568 s, System: 0.666 s]
  Range (min … max):   182.112 s … 182.488 s    5 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Summary
  after ran
    1.13 ± 0.08 times faster than before
```

</details>

- allowing cancellation via context to mitigate potential DoS attacks by immediately cancelling excessive requests