@xrgzs 2024-07-27
用 https://github.com/alist-org/alist/pull/6859 导致号被封禁了，估计是修改后调用接口超速了，请回滚到之前的版本

PS：之前的代码及限速部分为123官方人员提交，根据客服回应看，应该不会出现问题

![](https://github.com/user-attachments/assets/521dfd36-43d6-4b67-bec0-fc4560ca0c8a)

![](https://github.com/user-attachments/assets/3f66daec-00ff-4cf8-81c5-27533f13e369)

另外之前本人在使用时出现的性能问题为DNS服务器导致，现已查明

@seiuneko @xhofe