@seiuneko 2024-07-28
@xrgzs 经过对代码逻辑的确认，接口限流间隔依然是之前 123 云盘官方提交的 700 ms，下面是实际调试截图，可以看到是按 700 ms 间隔请求的：
![alist log](https://github.com/user-attachments/assets/10869902-6a02-434f-8d36-a96c7e1daaf4)

之前在测试时一天内多次触发 123 云盘风控，但账号能正常登录，只是 IP 被封禁，且光猫重新拨号后能正常使用：
![123pan web](https://github.com/user-attachments/assets/3fe96ca7-383b-4156-a1de-d78a9c680d08)

猜测账号封禁原因可能与你截图里显示的异地登录有关？